[MRISC32] Add a .clang-format file (revert me!)
The .clang-format file is a best-effort configuration that is meant to
aid in formatting the code according to the GNU code formatting
guidelines.

This overrides the upstream decision to ignore .clang-format files, and
it is expected that this commit needs to be reverted if the MRISC32 code
is ever upstreamed.